+++ apps/ui/src/components/ui/button.tsx
@@ -353,9 +353,9 @@ const buttonVariants = cva(
       },
       size: {
         default: "h-9 px-4 py-2",
+        xs: "h-6 px-2.5 py-1.5",
         sm: "h-8 px-3 py-1.5",
         lg: "h-10 px-8",
-        icon: "h-10 w-10",
         unsized: "",
       },
     },
